+++ .github/workflows/pr-check.yml
@@ -31,17 +31,6 @@ jobs:
             exit 0
           fi
 
-      - name: Check if CLI docs should be generated
-        id: check-paths
-        uses: dorny/paths-filter@v2
-        with:
-          filters: |
-            cli-docs:
-              - '.github/workflows/pr-check.yml'
-              - 'packages/cli/**'
-              - 'scripts/docs/package.json'
-              - 'scripts/docs/src/generate-cli-docs.ts'
-
       - name: Use Node.js
         uses: actions/setup-node@v2
         with:
@@ -83,12 +72,23 @@ jobs:
       - name: Test
         run: pnpm turbo test --force
 
-      - name: Require changeset to be present in PR
-        if: github.event.pull_request.user.login != 'dependabot[bot]'
-        run: pnpm changeset status --since origin/main
+      - name: Check if documentation should be generated
+        id: check-paths
+        uses: dorny/paths-filter@v2
+        with:
+          filters: |
+            cli-docs:
+              - '.github/workflows/pr-check.yml'
+              - 'packages/cli/**'
+              - 'scripts/docs/package.json'
+              - 'scripts/docs/src/generate-cli-docs.ts'
 
-      - name: Generate CLI docs
+      - name: Generate CLI documentation
         if: steps.check-paths.outputs.cli-docs == 'true'
         env:
           GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
         run: pnpm --filter docs run generate-cli-docs
+
+      - name: Require changeset to be present in PR
+        if: github.event.pull_request.user.login != 'dependabot[bot]'
+        run: pnpm changeset status --since origin/main
